Refactor gem to multiple modules, roughly divided by AWS service
