+++ lib/movable_ink/aws/errors.rb
@@ -1,0 +1,6 @@
+module MovableInk
+  class AWS
+    class FailedWithBackoff < StandardError; end
+    class EC2Required < StandardError; end
+  end
+end
